fix(karma): Add cli config poll option to karma default config (#2486)
With https://github.com/angular/angular-cli/pull/1814 the `defaults.poll` property was added to angular-cli.json. This configuration setting currently applies to `ng serve` but not `ng test`. This fix adds the poll value to the karma defaults so that if you set `defaults: { poll: 1000 }` it will apply to both `ng serve` and `ng test`.